Changed multiply output to half precision

+++ multiply/multiply.m
@@ -1,4 +1,3 @@
 function out = multiply(a,b)
-out = a * b;
+out = half(a * b);
 end
-
